+++ packages/google_maps_flutter/google_maps_flutter_web/lib/src/google_maps_controller.dart
@@ -7,7 +7,7 @@ part of '../google_maps_flutter_web.dart';
 /// Type used when passing an override to the _createMap function.
 @visibleForTesting
 typedef DebugCreateMapFunction =
-    gmaps.Map Function(HTMLElement div, gmaps.MapOptions options);
+    gmaps.Map Function(web.HTMLElement div, gmaps.MapOptions options);
 
 /// Type used when passing an override to the _setOptions function.
 @visibleForTesting
@@ -38,13 +38,58 @@ class GoogleMapController {
     _heatmapsController = HeatmapsController();
     _polygonsController = PolygonsController(stream: _streamController);
     _polylinesController = PolylinesController(stream: _streamController);
-    _clusterManagersController = ClusterManagersController(
-      stream: _streamController,
-    );
-    _markersController = MarkersController(
-      stream: _streamController,
-      clusterManagersController: _clusterManagersController!,
-    );
+
+    // Check if all markers are of the same type. Mixing marker types is not
+    // allowed.
+    final Set<Type> markerTypes = _markers
+        .map((Marker e) => e.runtimeType)
+        .toSet();
+    if (markerTypes.isNotEmpty) {
+      assert(markerTypes.length == 1, 'All markers must be of the same type.');
+
+      switch (mapConfiguration.markerType) {
+        case null:
+        case MarkerType.marker:
+          assert(
+            markerTypes.first == Marker,
+            'All markers must be of type Marker because '
+            'mapConfiguration.markerType is MarkerType.marker',
+          );
+        case MarkerType.advancedMarker:
+          assert(
+            markerTypes.first == AdvancedMarker,
+            'All markers must be of type AdvancedMarker because '
+            'mapConfiguration.markerType is MarkerType.advanced',
+          );
+      }
+    }
+
+    // Advanced and legacy markers are handled differently so markers controller
+    // and cluster manager need be initialized with the correct marker type.
+    _clusterManagersController = switch (mapConfiguration.markerType) {
+      null || MarkerType.marker => ClusterManagersController<gmaps.Marker>(
+        stream: _streamController,
+      ),
+      MarkerType.advancedMarker =>
+        ClusterManagersController<gmaps.AdvancedMarkerElement>(
+          stream: _streamController,
+        ),
+    };
+    _markersController = switch (mapConfiguration.markerType) {
+      null || MarkerType.marker => LegacyMarkersController(
+        stream: stream,
+        clusterManagersController:
+            _clusterManagersController!
+                as ClusterManagersController<gmaps.Marker>,
+      ),
+      MarkerType.advancedMarker => AdvancedMarkersController(
+        stream: stream,
+        clusterManagersController:
+            _clusterManagersController!
+                as ClusterManagersController<gmaps.AdvancedMarkerElement>,
+      ),
+    };
+
     _tileOverlaysController = TileOverlaysController();
     _groundOverlaysController = GroundOverlaysController(
       stream: _streamController,
@@ -98,7 +143,7 @@ class GoogleMapController {
 
   // The Flutter widget that contains the rendered Map.
   HtmlElementView? _widget;
-  late HTMLElement _div;
+  late web.HTMLElement _div;
 
   /// The Flutter widget that will contain the rendered Map. Used for caching.
   Widget? get widget {
@@ -133,8 +178,8 @@ class GoogleMapController {
   HeatmapsController? _heatmapsController;
   PolygonsController? _polygonsController;
   PolylinesController? _polylinesController;
-  MarkersController? _markersController;
-  ClusterManagersController? _clusterManagersController;
+  MarkersController<Object?, Object>? _markersController;
+  ClusterManagersController<Object?>? _clusterManagersController;
   TileOverlaysController? _tileOverlaysController;
   GroundOverlaysController? _groundOverlaysController;
 
@@ -151,7 +196,7 @@ class GoogleMapController {
 
   /// The ClusterManagersController of this Map. Only for integration testing.
   @visibleForTesting
-  ClusterManagersController? get clusterManagersController =>
+  ClusterManagersController<Object?>? get clusterManagersController =>
       _clusterManagersController;
 
   /// The GroundOverlaysController of this Map. Only for integration testing.
@@ -164,12 +209,12 @@ class GoogleMapController {
   void debugSetOverrides({
     DebugCreateMapFunction? createMap,
     DebugSetOptionsFunction? setOptions,
-    MarkersController? markers,
+    MarkersController<Object?, Object>? markers,
     CirclesController? circles,
     HeatmapsController? heatmaps,
     PolygonsController? polygons,
     PolylinesController? polylines,
-    ClusterManagersController? clusterManagers,
+    ClusterManagersController<Object?>? clusterManagers,
     TileOverlaysController? tileOverlays,
     GroundOverlaysController? groundOverlays,
   }) {
@@ -188,7 +233,7 @@ class GoogleMapController {
   DebugCreateMapFunction? _overrideCreateMap;
   DebugSetOptionsFunction? _overrideSetOptions;
 
-  gmaps.Map _createMap(HTMLElement div, gmaps.MapOptions options) {
+  gmaps.Map _createMap(web.HTMLElement div, gmaps.MapOptions options) {
     if (_overrideCreateMap != null) {
       return _overrideCreateMap!(div, options);
     }
@@ -467,7 +512,7 @@ class GoogleMapController {
         await Future<gmaps.LatLngBounds?>.value(_googleMap!.bounds) ??
         _nullGmapsLatLngBounds;
 
-    return gmLatLngBoundsTolatLngBounds(bounds);
+    return gmLatLngBoundsToLatLngBounds(bounds);
   }
 
   /// Returns the [ScreenCoordinate] for a given viewport [LatLng].
@@ -648,6 +693,13 @@ class GoogleMapController {
   /// Returns true if the [InfoWindow] of the marker identified by [MarkerId] is shown.
   bool isInfoWindowShown(MarkerId markerId) {
     return _markersController?.isInfoWindowShown(markerId) ?? false;
+  }
+
+  /// Returns true if this map supports [AdvancedMarker]s.
+  bool isAdvancedMarkersAvailable() {
+    assert(_googleMap != null, 'Cannot get map capabilities of a null map.');
+
+    return _googleMap!.mapCapabilities.isAdvancedMarkersAvailable ?? false;
   }
 
   // Cleanup