+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,6 +1,12 @@
-## 0.9.23+3
+## 0.10.0+1
 
 * Migrates to Swift-based pigeon interface.
+
+## 0.10.0
+
+* Adds video stabilization.
+  - Adds getSupportedVideoStabilizationModes method.
+  - Adds setVideoStabilizationMode method.
 
 ## 0.9.23+2